Create DateTimeFormatExample.java
In Java, when we declare a class as public, it's expected to be saved in a file with the same name as the class, including the capitalization. This is a naming convention enforced by Java.

So, if we have a class named DateTimeFormatExample declared as public, the file containing this class should be named DateTimeFormatExample.java.