A2sa4
We define a Student class with three data members: rollNumber, name, and percentage. We also have a static member objectCount to keep track of the number of objects created.

The Student class has both a default constructor and a parameterized constructor. Both constructors increment the objectCount when an object is created.

We create three student objects using the parameterized constructor and display the object count after each object is created. We also display the contents of each object using the display method.

Finally, we create a student object using the default constructor and display its contents.